Fixed picking bug Updated todo

+++ src/main/kotlin/com/pafoid/skate/engine/render/renderer/IRenderer.kt
@@ -7,7 +7,7 @@ import org.joml.Vector3f
 interface IRenderer {
     var useFbo: Boolean
     fun render(scene: Scene, activeGameObject: GameObject? = null, hoveredGameObject: GameObject? = null)
-    fun readPixel(x: Int, y: Int): Int
+    fun readPixel(x: Float, y: Float): Int
     fun clearColor(sky: Vector3f)
     fun destroy()
 }
